Specify number of bits explicitly for encode tests.
Previously, test inputs were generated based on the size of the 'morton' template type. For 32-bit tests, uint_fast32_t (which on some systems is of course larger than 32 bits) was problematic, since Encode functions configured to produce exactly 32-bit encodings were being asked to handle inputs which would overflow that range. For example, a 'MortonNDLutEncoder<3, 10, 5>' encoder is configured at compile-time to process 10 bits from each input field, and terminates early as an optimization beyond 10 bits. Providing inputs with beyond the first 10 bits set produces undefined results.

+++ test/libmorton_test.h
@@ -68,7 +68,7 @@ typedef decode_f_3D_wrapper<uint_fast32_t, uint_fast16_t> decode_3D_32_wrapper;
 
 template<std::size_t FieldCount>
 uint64_t split_by_n(uint64_t input, size_t bitsRemaining) {
-	return (bitsRemaining == 0) ? input : (split_by_n<FieldCount>(input >> 1, bitsRemaining - 1) << FieldCount) | (input & 1);
+	return (bitsRemaining == 0) ? input : (split_by_n<FieldCount>(input >> 1, bitsRemaining - 1) << FieldCount) | (input & (uint64_t)1);
 }
 
 template<std::size_t FieldCount>
